fingerprintdemosystem/requirements.txt
pyserial==3.5
pillow==10.3.0
opencv-python==4.9.0.80
tk==0.1.0